+++ src/test/java/us/ihmc/yoVariables/parameters/LongParameterTest.java
@@ -97,7 +97,7 @@ public class LongParameterTest
    public void testLoadFromString()
    {
 
-      for(long s = 100L * (long)Integer.MIN_VALUE; s < 100L * (long)Integer.MAX_VALUE; s+= (long)Integer.MAX_VALUE)
+      for(long s = 100L * Integer.MIN_VALUE; s < 100L * Integer.MAX_VALUE; s+= Integer.MAX_VALUE)
       { 
          YoVariableRegistry dummy = new YoVariableRegistry("dummy");
          LongParameter param = new LongParameter("test", dummy);
